Moved theme toggle to header
